+++ apps/meteor/ee/app/apps/bridges/scheduler.ts
@@ -5,7 +5,7 @@ import type { IProcessor, IOnetimeSchedule, IRecurringSchedule, IJobContext } fr
 import { StartupType } from '@rocket.chat/apps-engine/definition/scheduler';
 import { SchedulerBridge } from '@rocket.chat/apps-engine/server/bridges/SchedulerBridge';
 
-import type { AppServerOrchestrator } from '../../../../ee/server/apps/orchestrator';
+import type { AppServerOrchestrator } from '../../../server/apps/orchestrator';
 
 function _callProcessor(processor: IProcessor['processor']): (job: Job) => Promise<void> {
 	return (job) => {
