+++ apps/meteor/ee/app/apps/converterService.ts
@@ -5,7 +5,7 @@ import type { IVisitor } from '@rocket.chat/apps-engine/definition/livechat';
 import { ServiceClass } from '@rocket.chat/core-services';
 import type { IAppsConverterService } from '@rocket.chat/core-services';
 
-import type { AppServerOrchestrator } from './orchestrator';
+import type { AppServerOrchestrator } from '../../server/apps/orchestrator';
 import { OrchestratorFactory } from './orchestratorFactory';
 
 export class AppsConverterService extends ServiceClass implements IAppsConverterService {
